+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day02/p003/Solution454_02.java
@@ -1,8 +1,4 @@
-package com.noob.algorithm.daily.plan02.day02;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
+package com.noob.algorithm.daily.plan02.day02.p003;
 
 /**
  * 🟡 454 四数相加II - https://leetcode.cn/problems/4sum-ii/description/
