+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day02/p004/Solution028_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day02;
+package com.noob.algorithm.daily.plan02.day02.p004;
 
 /**
  * 🟢 028 找出字符串中第一个匹配项的下标 - https://leetcode.cn/problems/find-the-index-of-the-first-occurrence-in-a-string/description/
